@@ -1,4 +1,4 @@
-package be.pxl.auctions.dao;
+package be.pxl.auctions.repository;
 
 import be.pxl.auctions.model.Auction;
 import org.springframework.data.jpa.repository.JpaRepository;
