@@ -4,10 +4,7 @@ package be.pxl.auctions.rest;
 import be.pxl.auctions.rest.resource.UserCreateResource;
 import be.pxl.auctions.rest.resource.UserDTO;
 import be.pxl.auctions.service.UserService;
-import be.pxl.auctions.util.exception.DuplicateEmailException;
-import be.pxl.auctions.util.exception.InvalidDateException;
-import be.pxl.auctions.util.exception.InvalidEmailException;
-import be.pxl.auctions.util.exception.RequiredFieldException;
+import jakarta.validation.Valid;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -40,7 +37,7 @@ public class UserRest {
 
 
     @PostMapping
-    public UserDTO createUser(@RequestBody UserCreateResource userCreateResource) {
+    public UserDTO createUser(@RequestBody @Valid UserCreateResource userCreateResource) {
         return userService.createUser(userCreateResource);
     }
 }
