@@ -4,8 +4,6 @@ import be.pxl.auctions.util.exception.AuctionNotFoundException;
 import be.pxl.auctions.util.exception.DuplicateEmailException;
 import be.pxl.auctions.util.exception.InvalidBidException;
 import be.pxl.auctions.util.exception.InvalidDateException;
-import be.pxl.auctions.util.exception.InvalidEmailException;
-import be.pxl.auctions.util.exception.RequiredFieldException;
 import be.pxl.auctions.util.exception.UserNotFoundException;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -25,7 +23,7 @@ public class RestResponseEntityExceptionHandler extends ResponseEntityExceptionH
 				new HttpHeaders(), HttpStatus.NOT_FOUND, request);
 	}
 
-	@ExceptionHandler(value = { DuplicateEmailException.class, InvalidBidException.class, InvalidDateException.class, InvalidEmailException.class, RequiredFieldException.class })
+	@ExceptionHandler(value = { DuplicateEmailException.class, InvalidBidException.class, InvalidDateException.class })
 	protected ResponseEntity<Object> handleNotAcceptable(RuntimeException ex, WebRequest request) {
 		return handleExceptionInternal(ex, ex.getMessage(),
 				new HttpHeaders(), HttpStatus.NOT_ACCEPTABLE, request);
