@@ -1,7 +1,5 @@
 package be.pxl.auctions.service;
 
-import be.pxl.auctions.dao.AuctionRepository;
-import be.pxl.auctions.dao.UserDao;
 import be.pxl.auctions.model.Auction;
 import be.pxl.auctions.model.Bid;
 import be.pxl.auctions.model.User;
@@ -9,13 +7,12 @@ import be.pxl.auctions.model.builder.AuctionBuilder;
 import be.pxl.auctions.model.builder.BidBuilder;
 import be.pxl.auctions.model.builder.BidCreateResourceBuilder;
 import be.pxl.auctions.model.builder.UserBuilder;
+import be.pxl.auctions.repository.AuctionRepository;
+import be.pxl.auctions.repository.UserRepository;
 import be.pxl.auctions.rest.resource.BidCreateResource;
 import be.pxl.auctions.rest.resource.BidDTO;
 import be.pxl.auctions.util.exception.AuctionNotFoundException;
 import be.pxl.auctions.util.exception.InvalidBidException;
-import be.pxl.auctions.util.exception.InvalidDateException;
-import be.pxl.auctions.util.exception.InvalidEmailException;
-import be.pxl.auctions.util.exception.RequiredFieldException;
 import be.pxl.auctions.util.exception.UserNotFoundException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -40,7 +37,7 @@ public class BidServiceCreateBidTest {
     private static final String USER_EMAIL = "sophie@pxl.be";
 
     @Mock
-    private UserDao userDao;
+    private UserRepository userRepository;
     @Mock
     private AuctionRepository auctionRepository;
     @InjectMocks
@@ -55,29 +52,8 @@ public class BidServiceCreateBidTest {
     }
 
     @Test
-    void throwsRequiredFieldExceptionWhenBidHasNoEmail() {
-        bidCreateResource.setEmail("");
-        assertThrows(RequiredFieldException.class,
-                () -> bidService.createBid(bidCreateResource, AUCTION_ID));
-    }
-
-    @Test
-    void throwsInvalidEmailExceptionWhenBidHasInvalidEmail() {
-        bidCreateResource.setEmail("jenspanis.com");
-        assertThrows(InvalidEmailException.class,
-                () -> bidService.createBid(bidCreateResource, AUCTION_ID));
-    }
-
-    @Test
-    void throwsRequiredFieldExceptionWhenBidHasNoPrice() {
-        bidCreateResource.setPrice(null);
-        assertThrows(RequiredFieldException.class,
-                () -> bidService.createBid(bidCreateResource, AUCTION_ID));
-    }
-
-    @Test
     void throwsUserNotFoundExceptionWhenThereIsNoUserWithEmail() {
-        when(userDao.findUserByEmail(USER_EMAIL)).thenReturn(Optional.empty());
+        when(userRepository.findUserByEmail(USER_EMAIL)).thenReturn(Optional.empty());
 
         assertThrows(UserNotFoundException.class,
                 () -> bidService.createBid(bidCreateResource, AUCTION_ID));
@@ -86,7 +62,7 @@ public class BidServiceCreateBidTest {
     @Test
     void throwsAuctionNotFoundExceptionWhenThereIsNoAuctionWithAuctionId() {
         User user = UserBuilder.anUser().build();
-        when(userDao.findUserByEmail(USER_EMAIL)).thenReturn(Optional.of(user));
+        when(userRepository.findUserByEmail(USER_EMAIL)).thenReturn(Optional.of(user));
         when(auctionRepository.findById(AUCTION_ID)).thenReturn(Optional.empty());
 
         assertThrows(AuctionNotFoundException.class,
@@ -98,7 +74,7 @@ public class BidServiceCreateBidTest {
         bidCreateResource.setPrice(50.0);
 
         User user = UserBuilder.anUser().build();
-        when(userDao.findUserByEmail(USER_EMAIL)).thenReturn(Optional.of(user));
+        when(userRepository.findUserByEmail(USER_EMAIL)).thenReturn(Optional.of(user));
         Auction auction = AuctionBuilder.anAuction().build();
         when(auctionRepository.findById(AUCTION_ID)).thenReturn(Optional.of(auction));
 
@@ -114,7 +90,7 @@ public class BidServiceCreateBidTest {
         bidCreateResource.setPrice(150.0);
 
         User user = UserBuilder.anUser().build();
-        when(userDao.findUserByEmail(USER_EMAIL)).thenReturn(Optional.of(user));
+        when(userRepository.findUserByEmail(USER_EMAIL)).thenReturn(Optional.of(user));
         Auction auction = AuctionBuilder.anAuction().withEndDate(LocalDate.now().minusDays(5)).build();
         when(auctionRepository.findById(AUCTION_ID)).thenReturn(Optional.of(auction));
 
@@ -130,11 +106,11 @@ public class BidServiceCreateBidTest {
         bidCreateResource.setPrice(150.0);
 
         User user = UserBuilder.anUser().build();
-        when(userDao.findUserByEmail(USER_EMAIL)).thenReturn(Optional.of(user));
+        when(userRepository.findUserByEmail(USER_EMAIL)).thenReturn(Optional.of(user));
         Auction auction = AuctionBuilder.anAuction().withEndDate(LocalDate.now().plusDays(1)).build();
         when(auctionRepository.findById(AUCTION_ID)).thenReturn(Optional.of(auction));
 
-        Bid bid = BidBuilder.aBid().withUser(userDao.findUserByEmail(BidCreateResourceBuilder.EMAIL).get()).withAmount(100.0).build();
+        Bid bid = BidBuilder.aBid().withUser(userRepository.findUserByEmail(BidCreateResourceBuilder.EMAIL).get()).withAmount(100.0).build();
         auction.getBids().add(bid);
 
         assertThrows(InvalidBidException.class,
@@ -146,7 +122,7 @@ public class BidServiceCreateBidTest {
         bidCreateResource.setPrice(150.0);
 
         User user = UserBuilder.anUser().build();
-        when(userDao.findUserByEmail(USER_EMAIL)).thenReturn(Optional.of(user));
+        when(userRepository.findUserByEmail(USER_EMAIL)).thenReturn(Optional.of(user));
         Auction auction = AuctionBuilder.anAuction().withEndDate(LocalDate.now().plusDays(1)).build();
         when(auctionRepository.findById(AUCTION_ID)).thenReturn(Optional.of(auction));
 
