@@ -1,7 +1,7 @@
 package be.pxl.auctions.service;
 
-import be.pxl.auctions.dao.UserDao;
 import be.pxl.auctions.model.User;
+import be.pxl.auctions.repository.UserRepository;
 import be.pxl.auctions.rest.resource.UserDTO;
 import be.pxl.auctions.util.exception.UserNotFoundException;
 import org.junit.jupiter.api.BeforeEach;
@@ -9,7 +9,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.time.LocalDate;
@@ -26,7 +25,7 @@ public class UserServiceGetUserByIdTest {
 	private static final long USER_ID = 5L;
 
 	@Mock
-	private UserDao userDao;
+	private UserRepository userRepository;
 	@InjectMocks
 	private UserService userService;
 	private User user;
@@ -43,14 +42,14 @@ public class UserServiceGetUserByIdTest {
 
 	@Test
 	public void throwsUserNotFoundExceptionWhenNoUserWithGivenIdFound() {
-		when(userDao.findUserById(USER_ID)).thenReturn(Optional.empty());
+		when(userRepository.findById(USER_ID)).thenReturn(Optional.empty());
 
 		assertThrows(UserNotFoundException.class, () -> userService.getUserById(USER_ID));
 	}
 
 	@Test
 	public void returnsUserWhenUserFoundWithGivenId() {
-		when(userDao.findUserById(USER_ID)).thenReturn(Optional.of(user));
+		when(userRepository.findById(USER_ID)).thenReturn(Optional.of(user));
 		UserDTO userById = userService.getUserById(USER_ID);
 		assertNotNull(userById);
 		assertEquals(USER_ID, userById.getId());
